+++ unity/general/Src/UnitTest/wrap/AutoStaticCodeRegister.cs
@@ -6,6 +6,8 @@
         {
             jsEnv.AddLazyStaticWrapLoader(typeof(Puerts.UnitTest.OptionalParametersClass), Puerts_UnitTest_OptionalParametersClass_Wrap.GetRegisterInfo);
                 
+            jsEnv.AddLazyStaticWrapLoader(typeof(Puerts.UnitTest.WrapperGenTest), Puerts_UnitTest_WrapperGenTest_Wrap.GetRegisterInfo);
+                
         }
     }
 }
